Chap25: BST.java, Fix Bug of Delete Operation
The previous code cause nullpointerexception when the current's right
children is null. The general throught that the previous code follows
is correct, but the current's right child doesn't need to be compared.

The cases that jumps into this section of code are that the
current doesn't have left child, i.e., it only has one subtree (right
subtree). What really matters is that whether the current is parent's
left child or right child. If current is a left child, reconnects
current's right subtree to the left side of the parent, vice versa.