JavaEE: update persistence.xml for h2 DB
Configuration seems to be incorrect or something should have gone wrong, the webapp or wildfly choose to use H2 in-memory db rather than the MySQL DB. Persistence.xml is now updated, as this problem cannot be fixed. For in-memory H2 db to work, version has to be 2.1 as well for unknown reason.

modified:   bookstore-backend/src/main/resources/META-INF/persistence.xml